Refactor `Upload` to use `Upload.Result`
This serves the purposes of consolidating various callbacks, as well as of
having a separate “Cancelled” event. The UI code is not interested in the
cancellation events as it is the UI that initiates cancellation, however,
with the change the code might be easier to follow, and technically a
cancellation attempt is not guaranteed to result in an actual cancellation
due to the threaded code.